Double the timeout on first send
If server was working with another client, it must wait for the next
data bit or it might arrive later and interrupt the new client. But if
the data bit was already sent and lost to signal conflict, the server
might will wait until timeout before considering the client dropped and
continuing with the new client. Thus the new client must tolerate a
longer wait for the first bit.

Refactor send loop into a helper function to keep within line count
limits.